// internal/thread/models/thread_models.ts
export interface Thread {
    author?: string;
    authorId?: number;
    created?: string;
    forum?: string;
    forumId?: number;
    id?: number;
    message?: string;
    slug?: string;
    title?: string;
    votes?: number;
    coalesce?: string
    nickname?: string;
    // message?: string;
}